+++ noggin/representation/__init__.py
@@ -11,3 +11,27 @@ class Representation(object):
         if attr in self.raw:
             return self.raw[attr]
         return []
+    
+    def get_properties(self):
+        """
+        Returns a dict of all the properties defined above and their values
+        """
+        return {
+            key: getattr(self, key)
+            for key, value in self.__class__.__dict__.items()
+            if isinstance(value, property)
+        }
+
+    def diff_fields(self, other):
+        """
+        Compares two instances of the User class, and returns the properties
+        with values that are different
+        """
+        if not isinstance(other, self.__class__):
+            raise ValueError(f"Can't diff a {self.__class__.__name__} instance against a {other.__class__.__name__} instance")
+
+        return [
+            key
+            for key, value in self.get_properties().items()
+            if value != other.get_properties()[key]
+        ]
